Extract Prompt Studio style helper
